entirely remove Manager, add some comments
