--- ebiten-backend/manager.go
@@ -1,5 +1,0 @@
-package ebitenbackend
-
-type GetCursorFn func() (x, y float32)
-
-type Manager struct{}
